requirement.txt
absl-py==0.15.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==4.2.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
cycler==0.11.0
debugpy @ file:///D:/bld/debugpy_1660619187469/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
et-xmlfile==1.0.1
exceptiongroup==1.0.4
flatbuffers==1.12
fonttools==4.37.4
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
h11==0.14.0
h5py==3.7.0
idna==3.4
imbalanced-learn==0.6.2
imblearn==0.0
importlib-metadata==4.12.0
ipykernel @ file:///D:/bld/ipykernel_1666723449132/work
ipython @ file:///D:/bld/ipython_1651240762678/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
joblib==1.2.0
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1673615989977/work
jupyter_core @ file:///D:/bld/jupyter_core_1658332454038/work
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
numpy==1.19.5
oauthlib==3.2.1
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1673482170163/work
pandas==1.3.5
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.2.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1677600924538/work
protobuf==3.20.0
psutil @ file:///D:/bld/psutil_1666155596250/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1672682006896/work
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2022.4
pywin32==303
PyYAML==6.0
pyzmq @ file:///D:/bld/pyzmq_1663830622471/work
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.23.1
scipy==1.4.1
selenium==4.7.2
simplejson==3.18.0
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
TA-Lib @ file:///D:/bld/ta-lib_1655668061091/work
tensorboard==2.1.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==1.1.0
threadpoolctl==3.1.0
tornado @ file:///D:/bld/tornado_1656937938087/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
trio==0.22.0
trio-websocket==0.9.2
tushare==1.2.89
typing-extensions==3.7.4.3
urllib3==1.26.12
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
websocket-client==0.57.0
Werkzeug==2.2.2
wrapt==1.12.1
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.1
